backend/requirements.txt
# Django core
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3

# ASGI + WebSocket
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2

# Database
mysqlclient==2.2.6

# Authentication
djangorestframework-simplejwt==5.4.0
PyJWT==2.10.1

# Celery
celery[redis]==5.4.0
django-celery-beat==2.7.0
redis==5.2.1

# File handling
Pillow==11.1.0
python-magic==0.4.27

# Encryption
cryptography==44.0.0
PyNaCl==1.5.0

# Firebase push notifications
firebase-admin==6.6.0

# Translation
argostranslate>=1.9.6
langdetect>=1.0.9

# HTTP + parsing
requests==2.32.3
beautifulsoup4==4.12.3

# Calendar events
icalendar==6.1.0

# Password hashing
bcrypt==4.2.1

# Environment & DO compatibility
python-dotenv==1.0.1
django-environ>=0.11.2
django-storages[s3]>=1.14
boto3>=1.34
whitenoise>=6.6
django-redis>=5.4
sentry-sdk[django,celery]>=1.40

# Utilities
gunicorn==23.0.0
uvicorn[standard]==0.34.0
websockets==14.1
